# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dataplat API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/recipes": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "List Recipes",
        "description": "List all recipes owned by the authenticated user.\n\nReturns all recipes in newest-first order (created_at DESC, id DESC).\nNo pagination for MVP \u2014 recipe counts per user are expected to be small.\n``total`` is included in the response envelope for forward-compatibility:\na future paginated version can truncate ``items`` while keeping ``total``\naccurate without a breaking schema change.\n\nAuth required (F-008).",
        "operationId": "list_recipes_api_recipes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Create Recipe",
        "description": "Create a new recipe.\n\nCreates a recipe row in Postgres and returns the new record.\nDuplicate names return 409 (detected via the Postgres UNIQUE constraint\n``recipe_name_key``).  Auth required (F-008).\n\n``owner_id`` is set to the authenticated user's id.  ``definition`` is\nstored as-is (any JSON object); structural validation is deferred to\nsynthesis-time (F-082).",
        "operationId": "create_recipe_api_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Get Recipe",
        "description": "Return the full recipe record for the given id.\n\nOwner-scoping: combines ``id == ?`` AND ``owner_id == ?`` in one query so\nthat a non-existent id and an id owned by another user both return 404\n(no-enumeration-leak, mirrors get_source / list_sources_by_collection).\n\nReturns ``RecipeOut`` (all 7 fields including ``definition``).\n\nAuth required (F-008).",
        "operationId": "get_recipe_api_recipes__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "recipes"
        ],
        "summary": "Update Recipe",
        "description": "Update an existing recipe's definition (and optionally description).\n\nOwner-scoping: combines ``id == ?`` AND ``owner_id == ?`` in one query so\nthat a non-existent id and an id owned by another user both return 404\n(no-enumeration-leak, mirrors get_recipe).\n\nFreeze guard (invariant #3): if any dataset has been materialized from\nthis recipe (``dataset.recipe_id == id``), the update is rejected with 409.\nThe ``definition`` is the transformation contract; it must not change after\nany dataset has been produced from it.\n\n``definition`` is always replaced in full (required field).\n``description`` is only updated when explicitly present in the request body\n(uses Pydantic v2 ``model_fields_set`` \u2014 absence means \"leave unchanged\";\nnull means \"clear\"; string means \"update\").\n\n``updated_at`` is bumped app-side to a concrete UTC datetime (testability\npreference over ``func.now()`` \u2014 see agreed.md \u00a76).\n\nAuth required (F-008).",
        "operationId": "update_recipe_api_recipes__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{id}/preview": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Preview Recipe",
        "description": "Dry-run preview: fetch candidate chunks and synthesise LLM samples.\n\nOwner-scoping: combines ``id == ?`` AND ``owner_id == ?`` in one query so\nthat a non-existent id and an id owned by another user both return 404\n(no-enumeration-leak, mirrors get_recipe / update_recipe).\n\nSteps:\n  1. Load recipe (owner-scoped) \u2014 404 if not found or wrong owner.\n  2. Extract ``schema.template`` from ``recipe.definition``; 400 if absent.\n  3. Pass ``where_clause`` (from ``definition.filter.where``) and ``config``\n     (from ``definition.schema.config``) to ``run_preview``.\n  4. ``run_preview`` fetches up to ``n_samples`` chunks from Lance, calls\n     the LLM gateway once per chunk (``asyncio.gather``), and returns the\n     synthesised samples.\n  5. Errors from Lance (``LanceQueryError``) \u2192 400; errors from the preview\n     logic (``PreviewError``) \u2192 HTTP status code from the exception.\n\nNo writes to MinIO, Parquet, or any Postgres table are performed.\nAll intermediate data (chunks list, LLM responses) lives only in process\nmemory and is discarded after the response (invariant #2).\n\nAuth required (F-008).",
        "operationId": "preview_recipe_api_recipes__id__preview_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipePreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipePreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecipeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Definition"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "definition"
        ],
        "title": "RecipeCreate",
        "description": "Request body for POST /api/recipes."
      },
      "RecipeListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "RecipeListItem",
        "description": "Slim response schema for a single recipe in a list context.\n\nUsed by GET /api/recipes (F-038).  Omits ``owner_id`` (the endpoint is\nalready owner-scoped, so re-echoing it is redundant) and ``definition``\n(can be an arbitrarily large JSONB blob; not needed for list/browse use).\nFull record retrieval belongs on the detail endpoint (future F-039)."
      },
      "RecipeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecipeListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RecipeListResponse",
        "description": "Envelope for GET /api/recipes (F-038).\n\n``total`` is the count of ALL recipes owned by the caller.  Included for\nforward-compatibility: a future paginated version can return a subset in\n``items`` while keeping ``total`` accurate, without a breaking schema change."
      },
      "RecipeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Definition"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "owner_id",
          "definition",
          "created_at",
          "updated_at"
        ],
        "title": "RecipeOut",
        "description": "Response schema for a single recipe row.\n\nUsed by POST /api/recipes (F-037).  Returns the full record (id, name,\ndescription, owner_id, definition, created_at, updated_at) so that the\nclient has all fields in one round trip \u2014 consistent with SourceCollectionOut.\n\n``schema_template_operator_id`` is omitted for MVP; it will be added as an\noptional field in a later sprint when operator-linking semantics are defined."
      },
      "RecipePreviewRequest": {
        "properties": {
          "n_samples": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 3.0,
            "title": "N Samples",
            "default": 3
          }
        },
        "type": "object",
        "title": "RecipePreviewRequest",
        "description": "Request body for POST /api/recipes/{id}/preview (F-041).\n\n``n_samples`` controls how many candidate chunks are fetched from Lance and\npassed to the LLM gateway.  Pydantic enforces the [3, 5] range; a value\noutside that range returns 422 automatically (no custom handler needed)."
      },
      "RecipePreviewResponse": {
        "properties": {
          "samples": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "samples"
        ],
        "title": "RecipePreviewResponse",
        "description": "Response for POST /api/recipes/{id}/preview (F-041).\n\n``samples`` is a list of template-output dicts.  For ``sft_synthesis_qa``\neach item has at minimum ``\"instruction\"`` (str) and ``\"output\"`` (str).\nThe per-template output shape is intentionally loose (``dict[str, Any]``)\nat the API layer because future templates will emit different shapes; a\ndiscriminated union can be added in a later sprint without breaking this\ncontract.  There is no ``SamplesItem`` wrapper in MVP."
      },
      "RecipeUpdate": {
        "properties": {
          "definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Definition"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "definition"
        ],
        "title": "RecipeUpdate",
        "description": "Request body for PUT /api/recipes/{id} (F-040).\n\n``definition`` is REQUIRED \u2014 the whole definition dict is replaced (not merged).\nTyped as dict[str, Any] to reject non-object JSON (array, null, bare string) with\n422 at the Pydantic boundary, matching RecipeCreate semantics.\n\n``description`` is OPTIONAL \u2014 may be omitted to leave the existing description\nunchanged, or sent as null/None to clear it, or sent as a string to update it.\nThe handler uses ``body.model_fields_set`` to distinguish \"caller omitted the field\"\n(leave existing value) from \"caller explicitly sent null or a string\" (apply it).\nThis is the canonical Pydantic v2 idiom for partial-update fields \u2014 no non-serializable\nsentinel, no PydanticJsonSchemaWarning, clean ``default: null`` in the OpenAPI schema.\n\n``name`` is NOT present \u2014 renaming is deferred; MVP locks names after creation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}